chore: reorganized section blocks and renamed responsive columns to section columns
