+++ src/blocks/Section/Component.tsx
@@ -1,11 +1,11 @@
 import { SectionBlock as SectionBlockProps } from '@/payload-types';
 import { cn } from '@/utilities/cn';
-import { ResponsiveColumnsBlock } from '../ResponsiveColumns/Component';
-import { SectionTitleBlock } from '../SectionTitle/Component';
+import { SectionColumnsBlock } from './blocks/SectionColumns/Component';
+import { SectionTitleBlock } from './blocks/SectionTitle/Component';
 
 const sectionBlockComponents = {
   sectionTitle: SectionTitleBlock,
-  responsiveCols: ResponsiveColumnsBlock,
+  sectionCols: SectionColumnsBlock,
 };
 
 export const SectionBlock: React.FC<SectionBlockProps> = ({
