+++ src/blocks/Section/blocks/SectionTitle/config.ts
