+++ src/blocks/Section/blocks/index.ts
@@ -1,0 +1,4 @@
+import { SectionColumns } from './SectionColumns/config';
+import { SectionTitle } from './SectionTitle/config';
+
+export { SectionColumns, SectionTitle };
