+++ src/blocks/Section/index.ts
@@ -1,0 +1,9 @@
+import { SectionColumns } from './blocks/SectionColumns/config';
+import { SectionTitle } from './blocks/SectionTitle/config';
+import { Section as SectionRoot } from './config';
+
+export const Section = {
+  Root: SectionRoot,
+  Columns: SectionColumns,
+  Title: SectionTitle,
+};
