+++ src/blocks/index.ts
@@ -1,7 +1,6 @@
 import { HeroSpinner } from './HeroSpinner/config';
 import { HiddenTitle } from './HiddenTitle/config';
-import { ResponsiveColumns } from './ResponsiveColumns/config';
-import { Section } from './Section/config';
-import { SectionTitle } from './SectionTitle/config';
+import { Section } from './Section';
+import { SectionColumns, SectionTitle } from './Section/blocks';
 
-export { HeroSpinner, HiddenTitle, ResponsiveColumns, Section, SectionTitle };
+export { HeroSpinner, HiddenTitle, Section, SectionColumns, SectionTitle };
