+++ src/payload.config.ts
@@ -6,7 +6,7 @@ import path from 'path';
 import { buildConfig } from 'payload';
 import sharp from 'sharp';
 import { fileURLToPath } from 'url';
-import { HeroSpinner, HiddenTitle, ResponsiveColumns, Section, SectionTitle } from './blocks';
+import { HeroSpinner, HiddenTitle, Section } from './blocks';
 import { Media } from './collections/Media';
 import { Pages } from './collections/Pages';
 import { Users } from './collections/Users';
@@ -48,7 +48,7 @@ export default buildConfig({
   },
   collections: [Pages, Media, Users],
   globals: [Header, Footer],
-  blocks: [HeroSpinner, HiddenTitle, Section, SectionTitle, ResponsiveColumns],
+  blocks: [HeroSpinner, HiddenTitle, Section.Root, Section.Title, Section.Columns],
   editor: lexicalEditor(),
   secret: process.env.PAYLOAD_SECRET || '',
   typescript: {
